requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.3
cssselect2==0.7.0
Django==5.1.1
django-allauth==65.0.2
django-celery-beat==2.7.0
django-paypal==2.1
django-timezone-field==7.0
fonttools==4.55.0
idna==3.10
kombu==5.4.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.15.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
zopfli==0.2.3.post1
